# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "operationId": "AdminController_listUsers",
        "summary": "分页查询用户列表",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "当前页码，默认 1",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数（最大100），默认 20",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "搜索关键词（用户名/昵称/QQ）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortField",
            "required": false,
            "in": "query",
            "description": "排序字段：id/username/nickname/role/status/createdAt/lastLoginAt/loginCount/level/playerName/location/affinity/playTime",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "排序方向：asc 或 desc，默认 asc",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/users/update": {
      "post": {
        "operationId": "AdminController_updateUser",
        "summary": "更新用户角色/状态/昵称/QQ号",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/users/detail": {
      "post": {
        "operationId": "AdminController_getUserDetail",
        "summary": "获取用户详细信息(含玩家档案、在线状态、累计在线时长等)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDetailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/players/edit": {
      "post": {
        "operationId": "AdminController_editPlayerData",
        "summary": "批量编辑玩家游戏数据(字段白名单，数值自动校验)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditPlayerDataDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/users/delete": {
      "post": {
        "operationId": "AdminController_deleteUser",
        "summary": "删除用户(级联删除其玩家档案，不可删除自己/超级管理员)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/users/reset-data": {
      "post": {
        "operationId": "AdminController_resetUserData",
        "summary": "清理用户游戏数据(重置为未开始游玩，保留账号)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetUserDataDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/users/reset-home": {
      "post": {
        "operationId": "AdminController_resetUserHome",
        "summary": "重置用户家园(进度/家园名/动态图归零，等级背包等其它进度不动)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetUserDataDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/users/batch-reset-home": {
      "post": {
        "operationId": "AdminController_batchResetUserHome",
        "summary": "批量重置所选用户家园(仅家园维度)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUserIdsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/users/reset-all-homes": {
      "post": {
        "operationId": "AdminController_resetAllUserHomes",
        "summary": "一键重置全部玩家家园(仅家园维度，保留等级/背包/账号)",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/users/batch-delete": {
      "post": {
        "operationId": "AdminController_batchDeleteUsers",
        "summary": "批量删除用户账号(级联删除其玩家档案；自动跳过自己/超级管理员)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUserIdsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/users/batch-reset-data": {
      "post": {
        "operationId": "AdminController_batchResetUserData",
        "summary": "批量清理用户游戏数据(重置为未开始游玩，保留账号)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUserIdsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/users/reset-all-data": {
      "post": {
        "operationId": "AdminController_resetAllUserData",
        "summary": "一键清空全部玩家游戏数据(保留所有账号)",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/config": {
      "get": {
        "operationId": "AdminController_listConfig",
        "summary": "获取所有系统配置项",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/config/export": {
      "get": {
        "operationId": "AdminController_exportConfig",
        "summary": "导出系统配置为 JSON 文件",
        "description": "整表导出（含 type/group/label/description），供跨部署共享或测试库→正式库同步。未来新增配置项自动包含，无需改此接口。",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/config/import": {
      "post": {
        "operationId": "AdminController_importConfig",
        "summary": "导入系统配置包",
        "description": "payload 为 GET /admin/config/export 下载的 JSON。merge：按键合并（有则改、无则建）；replace：清空后全量写入。dryRun=true 时只返回差异统计不落库。",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportConfigDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/config/{key}": {
      "get": {
        "operationId": "AdminController_getConfig",
        "summary": "获取单个配置项",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/config/update": {
      "post": {
        "operationId": "AdminController_updateConfig",
        "summary": "更新系统配置项(在线生效)",
        "description": "修改后立即生效，无需重启；全部配置键见 GET /admin/config。签到奖励相关键：game.checkinBaseExp(签到基础经验) / game.checkinConsecutiveExpPerDay(每连续1天额外经验) / game.checkinConsecutiveExpMaxDays(连续加成封顶天数，0=不封顶) / game.checkinRewards(奖励表 JSON：{ dailyCycleDays, daily, consecutive, total }，奖励条目 { type: item|exp|vitality, name, quantity })。",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConfigDto"
              },
              "examples": {
                "number": {
                  "summary": "数值型示例（签到基础经验）",
                  "value": {
                    "key": "game.checkinBaseExp",
                    "value": 50
                  }
                },
                "checkinRewards": {
                  "summary": "签到奖励表示例（JSON 型）",
                  "description": "daily=按连续第 N 天发放(dailyCycleDays>0 时按周期轮转)；consecutive=连续签到里程碑；total=累计签到里程碑",
                  "value": {
                    "key": "game.checkinRewards",
                    "value": {
                      "dailyCycleDays": 7,
                      "daily": [
                        {
                          "days": 1,
                          "rewards": [
                            {
                              "type": "item",
                              "name": "签到礼包",
                              "quantity": 1
                            }
                          ]
                        }
                      ],
                      "consecutive": [
                        {
                          "days": 7,
                          "rewards": [
                            {
                              "type": "item",
                              "name": "签到礼包",
                              "quantity": 1
                            }
                          ]
                        }
                      ],
                      "total": [
                        {
                          "days": 30,
                          "rewards": [
                            {
                              "type": "item",
                              "name": "累计签到礼包",
                              "quantity": 1
                            }
                          ]
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/status": {
      "get": {
        "operationId": "AdminController_getServerStatus",
        "summary": "获取服务器状态（用户数、在线玩家、地图数等）",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/dashboard": {
      "get": {
        "operationId": "AdminController_getDashboard",
        "summary": "获取服务器仪表盘数据（同 status）",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/announcement": {
      "post": {
        "operationId": "AdminController_sendAnnouncement",
        "summary": "发送系统公告（广播给所有在线玩家，兼容 content/message 字段）",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/gm/announcement": {
      "post": {
        "operationId": "AdminController_sendGmAnnouncement",
        "summary": "发送系统公告（GM 兼容路径）",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/admin/players": {
      "get": {
        "operationId": "AdminController_getPlayers",
        "summary": "获取玩家列表（管理员用，分页）",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "管理员"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "使魔大战3 网页版 API",
    "description": "使魔大战3 网页版后端接口文档（OpenAPI 3.0 规范）",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "用户ID",
            "example": 1
          },
          "role": {
            "type": "string",
            "description": "新角色",
            "enum": [
              "USER",
              "ADMIN",
              "SUPER_ADMIN"
            ]
          },
          "status": {
            "type": "string",
            "description": "账号状态",
            "enum": [
              "ACTIVE",
              "BANNED"
            ]
          },
          "nickname": {
            "type": "string",
            "description": "昵称"
          },
          "qqNumber": {
            "type": "string",
            "description": "QQ号(可修改，解绑传空字符串)",
            "example": "123456789"
          }
        },
        "required": [
          "id"
        ]
      },
      "UserDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "用户ID",
            "example": 1
          }
        },
        "required": [
          "id"
        ]
      },
...
      "DeleteUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "用户ID",
            "example": 1
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}