# server/src/modules/admin/admin.controller.ts
/**
 * 管理员控制器
 * 提供用户管理、系统配置管理、游戏管理等后台管理接口，仅 ADMIN 及以上角色可访问。
 * 所有接口均生成 OpenAPI 文档。
 */

import {
  BadRequestException,
  Body,
  Controller,
  DefaultValuePipe,
  Delete,
  Get,
  Optional,
  Param,
  ParseIntPipe,
  Post,
  Put,
  Query,
  Req,
  UploadedFiles,
  UseGuards,
  UseInterceptors,
} from '@nestjs/common';
import { FilesInterceptor } from '@nestjs/platform-express';
import { diskStorage } from 'multer';
import { join } from 'path';
import { existsSync, mkdirSync } from 'fs';
import { buildUploadFilename } from '../../common/utils/upload-filename.util';
import { ApiBearerAuth, ApiBody, ApiConsumes, ApiOperation, ApiQuery, ApiTags } from '@nestjs/swagger';
import { Roles } from '../../common/decorators/roles.decorator';
import { RolesGuard } from '../../common/guards/roles.guard';
import { JwtAuthGuard } from '../auth/jwt-auth.guard';
import { GlobalConfig } from '../../config/global.config';
import { SystemConfigService } from '../system-config/system-config.service';
import { GlobalProficiencyService, GLOBAL_MARKERS_KEY } from '../game/global-proficiency.service';
import { AdminService } from './admin.service';
import { StaticDataAdminService } from './static-data-admin.service';
import {
  AnnouncementDto,
  BackpackSaveDto,
  BatchUserIdsDto,
  DeleteUserDto,
  EditPlayerDataDto,
  GameDataSaveDto,
  ResetUserDataDto,
  SetWorldLevelDto,
  ImportConfigDto,
  UpdateConfigDto,
  UpdateUserDto,
  UserDetailDto,
} from './dto/admin.dto';

@ApiTags('管理员')
@ApiBearerAuth()
@UseGuards(JwtAuthGuard, RolesGuard) // 先登录，再校验角色
@Roles('ADMIN', 'SUPER_ADMIN') // 整个控制器要求管理员及以上
@Controller('admin')
export class AdminController {
  constructor(
    private readonly adminService: AdminService,
    private readonly staticDataAdminService: StaticDataAdminService,
    private readonly systemConfigService: SystemConfigService,
    // GameModule 为 @Global；循环依赖场景下可选注入，缺失时仅跳过内存同步
    @Optional() private readonly globalProficiency?: GlobalProficiencyService,
  ) {}

  /// ===== 用户管理 =====

  @Get('users')
  @ApiOperation({ summary: '分页查询用户列表' })
  @ApiQuery({ name: 'page', required: false, description: '当前页码，默认 1', type: Number })
  @ApiQuery({ name: 'pageSize', required: false, description: '每页条数（最大100），默认 20', type: Number })
  @ApiQuery({ name: 'keyword', required: false, description: '搜索关键词（用户名/昵称/QQ）', type: String })
  @ApiQuery({
    name: 'sortField',
    required: false,
    description:
      '排序字段：id/username/nickname/role/status/createdAt/lastLoginAt/loginCount/level/playerName/location/affinity/playTime',
    type: String,
  })
  @ApiQuery({ name: 'sortOrder', required: false, description: '排序方向：asc 或 desc，默认 asc', type: String })
  async listUsers(
    @Query('page', new DefaultValuePipe(1), ParseIntPipe) page: number,
    @Query('pageSize', new DefaultValuePipe(20), ParseIntPipe) pageSize: number,
    @Query('keyword') keyword?: string,
    @Query('sortField') sortField?: string,
    @Query('sortOrder') sortOrder?: 'asc' | 'desc',
  ) {
    const data = await this.adminService.listUsers(page, pageSize, keyword, sortField, sortOrder);
    return { success: true, data };
  }

  @Post('users/update')
  @ApiOperation({ summary: '更新用户角色/状态/昵称/QQ号' })
  async updateUser(@Body() dto: UpdateUserDto) {
    const user = await this.adminService.updateUser(dto.id, {
      role: dto.role,
      status: dto.status,
      nickname: dto.nickname,
      qqNumber: dto.qqNumber,
    });
    return { success: true, data: user };
  }

  @Post('users/detail')
  @ApiOperation({ summary: '获取用户详细信息(含玩家档案、在线状态、累计在线时长等)' })
  async getUserDetail(@Body() dto: UserDetailDto) {
    const data = await this.adminService.getUserDetail(dto.id);
    return { success: true, data };
  }

  @Post('players/edit')
  @ApiOperation({ summary: '批量编辑玩家游戏数据(字段白名单，数值自动校验)' })
  async editPlayerData(@Body() dto: EditPlayerDataDto, @Req() req) {
    const message = await this.adminService.editPlayerData(
      req.user.userId,
      dto.id,
      dto.data,
    );
    return { success: true, message };
  }

  @Post('users/delete')
  @ApiOperation({ summary: '删除用户(级联删除其玩家档案，不可删除自己/超级管理员)' })
  async deleteUser(@Body() dto: DeleteUserDto, @Req() req) {
    const message = await this.adminService.deleteUser(req.user.userId, dto.id);
    return { success: true, message };
  }

  @Post('users/reset-data')
  @ApiOperation({ summary: '清理用户游戏数据(重置为未开始游玩，保留账号)' })
  async resetUserData(@Body() dto: ResetUserDataDto) {
    const message = await this.adminService.resetPlayerData(dto.id);
    return { success: true, message };
  }

  @Post('users/reset-home')
  @ApiOperation({ summary: '重置用户家园(进度/家园名/动态图归零，等级背包等其它进度不动)' })
  async resetUserHome(@Body() dto: ResetUserDataDto) {
    const message = await this.adminService.resetHomeData(dto.id);
    return { success: true, message };
  }

  @Post('users/batch-reset-home')
  @ApiOperation({ summary: '批量重置所选用户家园(仅家园维度)' })
  async batchResetUserHome(@Body() dto: BatchUserIdsDto) {
    const message = await this.adminService.batchResetHomeData(dto.ids);
    return { success: true, message };
  }

  @Post('users/reset-all-homes')
  @ApiOperation({ summary: '一键重置全部玩家家园(仅家园维度，保留等级/背包/账号)' })
  async resetAllUserHomes() {
    const message = await this.adminService.resetAllHomeData();
    return { success: true, message };
  }

  @Post('users/batch-delete')
  @ApiOperation({ summary: '批量删除用户账号(级联删除其玩家档案；自动跳过自己/超级管理员)' })
  async batchDeleteUsers(@Body() dto: BatchUserIdsDto, @Req() req) {
    const message = await this.adminService.batchDeleteUsers(req.user.userId, dto.ids);
    return { success: true, message };
  }

  @Post('users/batch-reset-data')
  @ApiOperation({ summary: '批量清理用户游戏数据(重置为未开始游玩，保留账号)' })
  async batchResetUserData(@Body() dto: BatchUserIdsDto) {
    const message = await this.adminService.batchResetPlayerData(dto.ids);
    return { success: true, message };
  }

  @Post('users/reset-all-data')
  @ApiOperation({ summary: '一键清空全部玩家游戏数据(保留所有账号)' })
  async resetAllUserData() {
    const message = await this.adminService.resetAllPlayerData();
    return { success: true, message };
  }

  /// ===== 系统配置中心 =====

  @Get('config')
  @ApiOperation({ summary: '获取所有系统配置项' })
  async listConfig() {
    const data = await this.systemConfigService.findAll();
    return { success: true, data };
  }

  /** 导出全部系统配置（须在 config/:key 之前注册，避免 export 被当成 key） */
  @Get('config/export')
  @ApiOperation({
    summary: '导出系统配置为 JSON 文件',
    description:
      '整表导出（含 type/group/label/description），供跨部署共享或测试库→正式库同步。' +
      '未来新增配置项自动包含，无需改此接口。',
  })
  async exportConfig() {
    const data = await this.systemConfigService.exportAll();
    return { success: true, data };
  }

  @Post('config/import')
  @ApiOperation({
    summary: '导入系统配置包',
    description:
      'payload 为 GET /admin/config/export 下载的 JSON。' +
      'merge：按键合并（有则改、无则建）；replace：清空后全量写入。' +
      'dryRun=true 时只返回差异统计不落库。',
  })
  async importConfig(@Body() dto: ImportConfigDto) {
    try {
      const result = await this.systemConfigService.importConfigs(dto.payload, {
        mode: dto.mode,
        dryRun: dto.dryRun === true,
      });
      // 全局熟练度有内存权威表；导入若覆盖该键必须同步，否则下次 flush 会覆盖导入值
      if (
        !result.dryRun &&
        this.globalProficiency &&
        Array.isArray(dto.payload?.configs) &&
        dto.payload.configs.some((c: any) => c?.key === GLOBAL_MARKERS_KEY)
      ) {
        const row = dto.payload.configs.find((c: any) => c?.key === GLOBAL_MARKERS_KEY);
        try {
          const raw = typeof row.value === 'string' ? JSON.parse(row.value || '{}') : row.value;
          let parsed: Record<string, number> = {};
          if (raw && typeof raw === 'object' && !Array.isArray(raw)) parsed = raw;
          await this.globalProficiency.replacePoints(parsed);
        } catch {
          // 坏数据已落库为原文；内存表保持不动
        }
      }
      const message = result.dryRun
        ? `预览完成：新增 ${result.created}、更新 ${result.updated}、无变化 ${result.unchanged}`
        : `导入完成：新增 ${result.created}、更新 ${result.updated}、无变化 ${result.unchanged}` +
          (result.deleted ? `、删除 ${result.deleted}` : '');
      return { success: true, message, data: result };
    } catch (err: any) {
      throw new BadRequestException(err?.message ?? '导入失败');
    }
  }

  @Get('config/:key')
  @ApiOperation({ summary: '获取单个配置项' })
  async getConfig(@Param('key') key: string) {
    const data = await this.systemConfigService.findByKey(key);
    return { success: true, data };
  }

  @Post('config/update')
  @ApiOperation({
    summary: '更新系统配置项(在线生效)',
    description:
      '修改后立即生效，无需重启；全部配置键见 GET /admin/config。签到奖励相关键：' +
      'game.checkinBaseExp(签到基础经验) / game.checkinConsecutiveExpPerDay(每连续1天额外经验) / ' +
      'game.checkinConsecutiveExpMaxDays(连续加成封顶天数，0=不封顶) / ' +
      'game.checkinRewards(奖励表 JSON：{ dailyCycleDays, daily, consecutive, total }，' +
      '奖励条目 { type: item|exp|vitality, name, quantity })。',
  })
  @ApiBody({
    type: UpdateConfigDto,
    examples: {
      number: {
        summary: '数值型示例（签到基础经验）',
        value: { key: 'game.checkinBaseExp', value: 50 },
      },
      checkinRewards: {
        summary: '签到奖励表示例（JSON 型）',
        description:
          'daily=按连续第 N 天发放(dailyCycleDays>0 时按周期轮转)；consecutive=连续签到里程碑；total=累计签到里程碑',
        value: {
          key: 'game.checkinRewards',
          value: {
            dailyCycleDays: 7,
            daily: [{ days: 1, rewards: [{ type: 'item', name: '签到礼包', quantity: 1 }] }],
            consecutive: [{ days: 7, rewards: [{ type: 'item', name: '签到礼包', quantity: 1 }] }],
            total: [{ days: 30, rewards: [{ type: 'item', name: '累计签到礼包', quantity: 1 }] }],
          },
        },
      },
    },
  })
  async updateConfig(@Body() dto: UpdateConfigDto) {
    const data = await this.systemConfigService.set(dto.key, dto.value);
    // 全局熟练度有内存权威表；整表替换后必须同步，否则下次 flush 会覆盖管理端修改
    if (dto.key === GLOBAL_MARKERS_KEY && this.globalProficiency) {
      let parsed: Record<string, number> = {};
      try {
        const raw = typeof dto.value === 'string' ? JSON.parse(dto.value) : dto.value;
        if (raw && typeof raw === 'object' && !Array.isArray(raw)) parsed = raw;
      } catch {
        // 非法 JSON 已在 set 落库为原文；内存表保持不动，避免被坏数据污染
      }
      if (Object.keys(parsed).length || dto.value === '{}' || dto.value === '') {
        await this.globalProficiency.replacePoints(parsed);
      }
    }
    return { success: true, data };
  }

  /// ===== 游戏管理 =====

  @Get('status')
  @ApiOperation({ summary: '获取服务器状态（用户数、在线玩家、地图数等）' })
  async getServerStatus() {
    const data = await this.adminService.getServerStatus();
    return { success: true, data };
  }

  /** 前端兼容：/admin/dashboard 映射到 /admin/status */
  @Get('dashboard')
  @ApiOperation({ summary: '获取服务器仪表盘数据（同 status）' })
  async getDashboard() {
    const data = await this.adminService.getServerStatus();
    return { success: true, data };
  }

  @Post('announcement')
  @ApiOperation({ summary: '发送系统公告（广播给所有在线玩家，兼容 content/message 字段）' })
  async sendAnnouncement(@Body() dto: AnnouncementDto) {
    const content = dto.content ?? dto.message;
    if (!content || !content.trim()) {
      throw new BadRequestException('公告内容不能为空');
    }
    await this.adminService.sendAnnouncement(content);
    return { success: true, message: '公告已发送' };
  }

  /** 前端兼容：/admin/gm/announcement 映射到 /admin/announcement */
  @Post('gm/announcement')
  @ApiOperation({ summary: '发送系统公告（GM 兼容路径）' })
  async sendGmAnnouncement(@Body() dto: AnnouncementDto) {
    const content = dto.content ?? dto.message;
    if (!content || !content.trim()) {
      throw new BadRequestException('公告内容不能为空');
    }
    await this.adminService.sendAnnouncement(content);
    return { success: true, message: '公告已发送' };
  }

  /**
   * 公告配图上传（仅图片）
   * 存储到 {uploadDir}/announcement/，返回可嵌入公告正文的相对 URL，
   * GM 编辑器以 Markdown 图片语法 ![描述](url) 插入正文。
   */
  @Post('announcement/upload')
  @ApiOperation({ summary: '上传公告配图（仅图片），返回可访问 URL 列表' })
  @ApiConsumes('multipart/form-data')
  @ApiBody({
    schema: { type: 'object', properties: { files: { type: 'array', items: { type: 'string', format: 'binary' } } } },
  })
  @UseInterceptors(
    FilesInterceptor('files', 5, {
      storage: diskStorage({
        destination: (_req, _file, cb) => {
          // 上传目录：{uploadDir}/announcement，不存在则创建
          const dir = join(process.cwd(), GlobalConfig.getInstance().uploadDir, 'announcement');
          if (!existsSync(dir)) {
            mkdirSync(dir, { recursive: true });
          }
          cb(null, dir);
        },
        filename: (_req, file, cb) => {
          // 文件名：时间戳 + 随机串 + 扩展名；剪贴板粘贴的图片无扩展名时按 MIME 兜底
          cb(null, buildUploadFilename(file.originalname, file.mimetype));
        },
      }),
      limits: {
        fileSize: GlobalConfig.getInstance().uploadMaxSize,
        files: 5,
      },
      fileFilter: (_req, file, cb) => {
        // 仅允许图片类型
        if (file.mimetype && file.mimetype.startsWith('image/')) {
          cb(null, true);
        } else {
          cb(new BadRequestException('公告配图仅支持图片文件'), false);
        }
      },
    }),
  )
  async uploadAnnouncementImages(@UploadedFiles() files: Express.Multer.File[]) {
    if (!files || files.length === 0) {
      return { success: false, message: '未收到文件' };
    }
    const prefix = GlobalConfig.getInstance().uploadUrlPrefix;
    // 返回相对访问路径，如 /uploads/announcement/xxx.png
    const urls = files.map((f) => `${prefix}/announcement/${f.filename}`);
    return { success: true, data: urls };
  }

  @Post('world-level')
  @ApiOperation({ summary: '设置世界等级（影响怪物强度和掉落）' })
  async setWorldLevel(@Body() dto: SetWorldLevelDto) {
    const message = await this.adminService.setWorldLevel(dto.level);
    return { success: true, message };
  }

  /** 前端兼容：获取世界等级 */
  @Get('gm/world-level')
  @ApiOperation({ summary: '获取当前世界等级' })
  async getWorldLevel() {
    const level = await this.adminService.getWorldLevel();
    return { success: true, data: { level } };
  }

  /** 前端兼容：/admin/gm/world-level 映射到 /admin/world-level */
  @Post('gm/world-level')
  @ApiOperation({ summary: '设置世界等级（GM 兼容路径）' })
  async setGmWorldLevel(@Body() dto: SetWorldLevelDto) {
    const message = await this.adminService.setWorldLevel(dto.level);
    return { success: true, message };
  }

  @Get('configs')
  @ApiOperation({ summary: '获取系统配置列表（可按分组筛选）' })
  async getConfigs(@Query('group') group?: string) {
    const data = await this.adminService.getSystemConfigs(group);
    return { success: true, data };
  }

  @Put('configs')
  @ApiOperation({ summary: '更新系统配置' })
  async updateAdminConfig(@Body() dto: UpdateConfigDto) {
    const message = await this.adminService.updateSystemConfig(dto.key, dto.value);
    return { success: true, message };
  }

  @Get('players')
  @ApiOperation({ summary: '获取玩家列表（管理员用，分页）' })
  async getPlayers(
    @Query('page') page = 1,
    @Query('pageSize') pageSize = 20,
  ) {
    const data = await this.adminService.getPlayersList(Number(page), Number(pageSize));
    return { success: true, data };
  }

  @Post('players/:id/ban')
  @ApiOperation({ summary: '封禁/解封用户（切换 ACTIVE ↔ BANNED）' })
  async toggleBan(@Param('id', ParseIntPipe) id: number) {
    const message = await this.adminService.toggleUserBan(id);
    return { success: true, message };
  }

  /** GM 物品目录（可发放的物品/装备名列表，供后台背包管理选择器与名称校验） */
  @Get('gm/catalog')
  @ApiOperation({ summary: 'GM 物品目录（物品+装备名称列表）' })
  getGmCatalog() {
    return { success: true, items: this.adminService.getGmItemCatalog() };
  }

  /** GM 背包管理：读取指定玩家背包的所有物品（完整解析） */
  @Get('gm/backpack')
  @ApiOperation({ summary: 'GM 读取玩家背包（解析背包全部物品）' })
  @ApiQuery({ name: 'userId', required: true, description: '目标用户ID', type: Number })
  async getBackpack(@Query('userId', ParseIntPipe) userId: number) {
    const data = await this.adminService.gmGetBackpack(userId);
    return { success: true, data };
  }

  /** GM 背包管理：整体保存背包（编辑数量 / 新增 / 删除） */
  @Post('gm/backpack/save')
  @ApiOperation({ summary: 'GM 保存玩家背包（编辑数量/增删物品，数量为0即删除）' })
  async saveBackpack(@Body() dto: BackpackSaveDto) {
    const message = await this.adminService.gmSaveBackpack(dto.userId, dto.items);
    return { success: true, message };
  }

  /// ===== 静态游戏数据管理（物品/装备/怪物/地图/任务等 JSON 配置） =====

  @Get('gamedata')
  @ApiOperation({ summary: '静态数据分类列表(含条目数/展示列/新增模板,驱动前端数据管理面板)' })
  listGameDataCategories() {
    return { success: true, data: this.staticDataAdminService.listCategories() };
  }

  @Get('gamedata/:key')
  @ApiOperation({ summary: '获取某分类全部条目(数据量小整包返回,前端本地过滤)' })
  async getGameDataEntries(@Param('key') key: string) {
    const data = await this.staticDataAdminService.getEntries(key);
    return { success: true, data };
  }

  @Post('gamedata/:key')
  @ApiOperation({ summary: '新增静态数据条目(写 JSON 文件+自动备份+热重载生效)' })
  async createGameDataEntry(@Param('key') key: string, @Body() dto: GameDataSaveDto) {
    const data = await this.staticDataAdminService.createEntry(key, dto.data);
    return { success: true, data };
  }

  @Put('gamedata/:key/:index')
  @ApiOperation({ summary: '更新静态数据条目(按下标,expectName 乐观校验)' })
  async updateGameDataEntry(
    @Param('key') key: string,
    @Param('index', ParseIntPipe) index: number,
    @Body() dto: GameDataSaveDto,
  ) {
    const data = await this.staticDataAdminService.updateEntry(key, index, dto.data, dto.expectName);
    return { success: true, data };
  }

  @Delete('gamedata/:key/:index')
  @ApiOperation({ summary: '删除静态数据条目(按下标,expectName 乐观校验;单配置分类禁删)' })
  async deleteGameDataEntry(
    @Param('key') key: string,
    @Param('index', ParseIntPipe) index: number,
    @Query('expectName') expectName?: string,
  ) {
    const data = await this.staticDataAdminService.deleteEntry(key, index, expectName);
    return { success: true, data };
  }

  }